+++ src/main/java/models/Candy.java
@@ -5,7 +5,7 @@ import com.techelevator.application.Inventory;
 import java.math.BigDecimal;
 
 public class Candy extends Inventory {
-    public Candy(String id, String name, BigDecimal price) {
+    public Candy(String id, String name, double price) {
         super(id, name, price);
     }
 }
