+++ src/main/java/models/Drink.java
@@ -5,7 +5,7 @@ import com.techelevator.application.Inventory;
 import java.math.BigDecimal;
 
 public class Drink extends Inventory {
-    public Drink(String id, String name, BigDecimal price) {
+    public Drink(String id, String name, double price) {
         super(id, name, price);
     }
 }
