+++ src/main/java/models/Munchy.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 public class Munchy extends Inventory {
 
-    public Munchy(String id, String name, BigDecimal price) {
+    public Munchy(String id, String name, double price) {
         super(id, name, price);
     }
